Move `read2` module to `utils` module
